+++ rfid_control.c
@@ -213,7 +213,7 @@ void read_all_rfid(){
     set_flip(1);
     protocol_buffer_synchronize();
     set_rfid(0);
-    for (size_t i = 0; i < TOOL_NUM-1; i++)
+    for (uint8_t i = 0; i < TOOL_NUM-1; i++)
     {
         memset(return_data, 0, 8);
         // 移动刀位置
